Cambiado tipo de dato de formatExtentDimensions a long
git-svn-id: http://svn.arte-consultores.com/svn/arte/istac/metamac/metamac-statistical-resources/trunk/metamac-statistical-resources@48499 05f23f4b-1f0b-0410-ba1b-ad9509270e59

+++ metamac-statistical-resources-rest-external-impl/src/test/java/org/siemac/metamac/rest/structural_resources/v1_0/utils/RestDoMocks.java
@@ -76,7 +76,7 @@ public class RestDoMocks {
         target.addStatisticalUnit(StatisticalResourcesDoMocks.mockConceptExternalItem("statisticalUnit2"));
         target.addStatisticalUnit(StatisticalResourcesDoMocks.mockConceptExternalItem("statisticalUnit3"));
         target.setRelatedDsd(StatisticalResourcesDoMocks.mockDsdExternalItem("DSD01"));
-        target.setFormatExtentObservations(Integer.valueOf(3));
+        target.setFormatExtentObservations(Long.valueOf(3));
         target.setFormatExtentDimensions(Integer.valueOf(5));
         target.setDateNextUpdate(new DateTime(2013, 12, 2, 3, 4, 5, 0));
         target.setUpdateFrequency(StatisticalResourcesDoMocks.mockCodeExternalItem("updateFrequency01"));
